fix: Support implict type conversion in Compiler (#432)
Committed-by: Xiaoli Zhou from Dev container

<!--
Thanks for your contribution! please review
https://github.com/GraphScope/neug/blob/main/CONTRIBUTING.md before
opening an issue.
-->

## What do these changes do?

as titled.

<!-- Please give a short brief about these changes. -->

## Related issue number

<!-- Are there any issues opened that will be resolved by merging this
change? -->

Fixes #416

---------

+++ tools/python_bind/tests/test_db_query.py
@@ -739,7 +739,7 @@ def test_complex_example(tmp_path):
         CREATE NODE TABLE Person(
             id INT64 PRIMARY KEY,
             name STRING,
-            age INT64,
+            age INT32,
             email STRING
         )
     """
@@ -751,7 +751,7 @@ def test_complex_example(tmp_path):
             id INT64 PRIMARY KEY,
             name STRING,
             industry STRING,
-            founded_year INT64
+            founded_year INT32
         )
     """
     )
@@ -772,7 +772,7 @@ def test_complex_example(tmp_path):
         """
         CREATE REL TABLE KNOWS(
             FROM Person TO Person,
-            since_year INT64,
+            since_year INT32,
             relationship_type STRING
         )
     """
